requirements.txt
Flask==2.0.1
PyJWT==2.3.0
pandas==1.3.3
psycopg2==2.9.1
streamlit==1.14.0
